Streamlit App/requirements.txt
streamlit
opencv-python
transformers
torch
Pillow